Fixed All memory leaks

+++ cAssetManager.cpp
@@ -40,7 +40,7 @@ bool cAssetManager::LoadTexture(std::string key, const char* filepath, SDL_Rende
 void cAssetManager::Draw(std::string key, int x, int y, int width, int height, float scrollingSpeed, int size, SDL_RendererFlip flip)
 {
 	cVector2 cameraPos = cCamera::Instance()->GetPosition() * scrollingSpeed;
-	SDL_Rect destinationRect = { x - cameraPos.m_x, y - cameraPos.m_y, width * size, height * size };
+	SDL_Rect destinationRect = { static_cast<int>(x - cameraPos.m_x), static_cast<int>(y - cameraPos.m_y), width * size, height * size };
 	
 	if (cGame::Instance()->GetRenderer() != nullptr && m_textureDictionary[key] != nullptr)
 	{
@@ -53,7 +53,7 @@ void cAssetManager::DrawAnimation(std::string key, int x, int y, int width, int 
 {
 	SDL_Rect sourceRect = { width * frame, height * row, width, height };
 	cVector2 cameraPos = cCamera::Instance()->GetPosition();
-	SDL_Rect destinationRect = { x - cameraPos.m_x, y - cameraPos.m_y, width * size, height * size };
+	SDL_Rect destinationRect = { static_cast<int>(x - cameraPos.m_x), static_cast<int>(y - cameraPos.m_y), width * size, height * size };
 
 	if (cGame::Instance()->GetRenderer() != nullptr && m_textureDictionary[key] != nullptr)
 	{
